"Fixed" Unreal Linking Bug
Galavant-Unreal was having a problem linking related to the STL. As it turns out, Unreal is linking against libc++, which is [broken on Ubuntu 16.04](http://stackoverflow.com/questions/37096062/get-a-basic-c-program-to-compile-using-clang-on-ubuntu-16/38385967#38385967). I changed the Jamrules so that it now links to libc++ (after the fiasco which was installing and hacking into libc++ locally).
Unfortunately, this broke linking my test projects because libc++ (or at least my local libc++) [doesn't have full STL support](http://stackoverflow.com/questions/7016730/compiling-with-clang-using-libc-undefined-references). This means editing the Jamrules every time you want to switch from compiling libGala* for Unreal to building unit tests. I'll eventually get this sorted out.